requirements.txt
attrs==19.3.0
click==7.1.1
Flask==1.1.1
importlib-metadata==1.5.0
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
marshmallow==3.5.1
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.6
pytest==5.3.5
python-dotenv==0.12.0
pytz==2019.3
six==1.14.0
wcwidth==0.1.8
Werkzeug==1.0.0
zipp==3.1.0
